[Refactor] AlcoholBox and FloatingReviewBtn components to utilize useReviewWrite hook for review handling
